+++ beacon-chain/chaintest/backend/simulated_backend.go
@@ -54,9 +54,8 @@ func NewSimulatedBackend() (*SimulatedBackend, error) {
 		return nil, fmt.Errorf("could not setup simulated backend db: %v", err)
 	}
 	cs, err := blockchain.NewChainService(context.Background(), &blockchain.Config{
-		BeaconDB:         db,
-		IncomingBlockBuf: 0,
-		EnablePOWChain:   false,
+		BeaconDB:       db,
+		EnablePOWChain: false,
 	})
 	if err != nil {
 		return nil, err
