+++ beacon-chain/rpc/validator_server_test.go
@@ -517,55 +517,3 @@ func TestWaitForActivation_ValidatorOriginallyExists(t *testing.T) {
 		t.Fatalf("Could not setup wait for activation stream: %v", err)
 	}
 }
-
-func TestWaitForActivation_ListensAndFetchesValidatorFromStateFeed(t *testing.T) {
-	db := internal.SetupDB(t)
-	defer internal.TeardownDB(t, db)
-
-	pubKey := []byte{'A'}
-
-	beaconState := &pbp2p.BeaconState{
-		Slot: params.BeaconConfig().GenesisSlot,
-		ValidatorRegistry: []*pbp2p.Validator{{
-			ActivationEpoch: params.BeaconConfig().GenesisSlot,
-			ExitEpoch:       params.BeaconConfig().FarFutureEpoch,
-			Pubkey:          pubKey},
-		},
-	}
-	if err := db.SaveState(beaconState); err != nil {
-		t.Fatalf("could not save state: %v", err)
-	}
-
-	vs := &ValidatorServer{
-		beaconDB:           db,
-		ctx:                context.Background(),
-		chainService:       newMockChainService(),
-		canonicalStateChan: make(chan *pbp2p.BeaconState, 1),
-	}
-	req := &pb.ValidatorActivationRequest{
-		Pubkey: pubKey,
-	}
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	mockStream := internal.NewMockValidatorService_WaitForActivationServer(ctrl)
-	mockStream.EXPECT().Send(
-		&pb.ValidatorActivationResponse{
-			Validator: beaconState.ValidatorRegistry[0],
-		},
-	).Return(nil)
-
-	exitRoutine := make(chan bool)
-	go func(tt *testing.T) {
-		if err := vs.WaitForActivation(req, mockStream); err != nil {
-			t.Fatalf("Could not setup wait for activation stream: %v", err)
-		}
-		<-exitRoutine
-	}(t)
-	vs.canonicalStateChan <- beaconState
-	if err := db.SaveValidatorIndex(pubKey, 0); err != nil {
-		t.Fatalf("Could not save validator index: %v", err)
-	}
-	vs.canonicalStateChan <- beaconState
-	exitRoutine <- true
-}
